studymate/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
supabase==2.4.6
PyMuPDF==1.24.5
google-generativeai==0.7.2
PyJWT==2.8.0
bcrypt==4.1.3
werkzeug==3.0.3
